+++ src/axom/core/execution/internal/cuda_exec.hpp
@@ -63,7 +63,7 @@ struct execution_space<CUDA_EXEC<BLOCK_SIZE, SYNCHRONOUS>>
   static constexpr char* name() noexcept { return (char*)"[CUDA_EXEC]"; }
   static int allocatorID() noexcept
   {
-    return axom::getUmpireResourceAllocatorID(umpire::resource::Unified);
+    return axom::getUmpireResourceAllocatorID(umpire::resource::Device);
   }
 };
 
@@ -93,7 +93,7 @@ struct execution_space<CUDA_EXEC<BLOCK_SIZE, ASYNC>>
   }
   static int allocatorID() noexcept
   {
-    return axom::getUmpireResourceAllocatorID(umpire::resource::Unified);
+    return axom::getUmpireResourceAllocatorID(umpire::resource::Device);
   }
 };
 }  // namespace axom