+++ src/axom/core/execution/internal/hip_exec.hpp
@@ -61,7 +61,7 @@ struct execution_space<HIP_EXEC<BLOCK_SIZE, SYNCHRONOUS>>
   static constexpr char* name() noexcept { return (char*)"[HIP_EXEC]"; }
   static int allocatorID() noexcept
   {
-    return axom::getUmpireResourceAllocatorID(umpire::resource::Unified);
+    return axom::getUmpireResourceAllocatorID(umpire::resource::Device);
   }
 };
 
@@ -88,7 +88,7 @@ struct execution_space<HIP_EXEC<BLOCK_SIZE, ASYNC>>
   static constexpr char* name() noexcept { return (char*)"[HIP_EXEC] (async)"; }
   static int allocatorID() noexcept
   {
-    return axom::getUmpireResourceAllocatorID(umpire::resource::Unified);
+    return axom::getUmpireResourceAllocatorID(umpire::resource::Device);
   }
 };
 }  // namespace axom